Use new React 16.3 context API to allow disabling fetching
Upgrade to React 16.3 to make use of the new public context API (https://reactjs.org/docs/context.html).

fetch-on-update now exports a component DisableFetchOnUpdate which allows disabling the behavior for all child components of wherever that component is used. The use case for this is server rendering. After loading data on a first-pass render, we don't want to trigger any more loads when rendering on the second pass.

I had to disable the tests for fetch-on-update for now until airbnb makes a new enzyme release including support for the new context node types. See airbnb/enzyme#1513.

+semver:major for more restrictive React version requirements now